Update 2015-02-03-how-betty-who-is-89-gets-her-news.md
From an overall design process strong consideration needs to be given to the functional reason as to why someone who didn't grow up with with technology would decide to use it. It could be as simple as a need to communicate with your grandchildren who live far away, or it could be as complex as finding ways to retain your independence and being self-sufficient after moving into a retirement residence. This question needs to be asked before venturing down the design-path, with the question "how much closer does this service get the user to fulfilling their need".